+++ prboom2/src/gl_preprocess.c
@@ -670,7 +670,7 @@ static void gld_PrecalculateSector(int num)
             // the angle of the last candidate
             // e6y: for finding an angle between AB and BC vectors we should subtract
             // (BC - BA) == (BC - (180 - AB)) == (angle-(180-lineangle))
-            if (D_abs(angle-(180-lineangle))<D_abs(bestangle))
+            if (D_abs((int) angle - (180 - (int) lineangle))<D_abs((int) bestangle))
             {
               bestline=i;
               bestangle=angle-(180-lineangle);
